+++ packages/spruce-event-plugin/src/.spruce/errors/options.types.ts
@@ -1,11 +1,10 @@
 import { SpruceErrors } from "#spruce/errors/errors.types"
-import { SpruceErrorOptions, ErrorOptions as ISpruceErrorOptions} from "@sprucelabs/error"
-import { SchemaErrorOptions } from '@sprucelabs/schema'
+import { ErrorOptions as ISpruceErrorOptions} from "@sprucelabs/error"
 
 export interface EventPluginErrorErrorOptions extends SpruceErrors.Event.EventPluginError, ISpruceErrorOptions {
 	code: 'EVENT_PLUGIN_ERROR'
 }
 
-type ErrorOptions = SchemaErrorOptions | SpruceErrorOptions | EventPluginErrorErrorOptions 
+type ErrorOptions =  | EventPluginErrorErrorOptions 
 
 export default ErrorOptions