+++ packages/spruce-event-plugin/src/plugins/event.plugin.ts
@@ -5,6 +5,7 @@ import {
 	SkillEventContract,
 	SpruceSchemas,
 } from '@sprucelabs/mercury-types'
+import { SchemaError } from '@sprucelabs/schema'
 import {
 	eventContractUtil,
 	eventDiskUtil,
@@ -183,7 +184,7 @@ export class EventFeaturePlugin implements SkillFeature {
 					await this.queueDidBoot(didBoot)
 				}
 			}
-		} catch (err) {
+		} catch (err: any) {
 			rej(err)
 			this._isBooted = false
 			this.isExecuting = false
@@ -203,7 +204,7 @@ export class EventFeaturePlugin implements SkillFeature {
 			const event = await this.buildSpruceEvent('did-boot')
 
 			await didBoot(event)
-		} catch (err) {
+		} catch (err: any) {
 			if (!this.executeReject) {
 				throw err
 			}
@@ -244,7 +245,7 @@ export class EventFeaturePlugin implements SkillFeature {
 			await this.destroy()
 
 			return health
-		} catch (err) {
+		} catch (err: any) {
 			const health: HealthCheckItem = {
 				status: 'failed',
 				errors: [
@@ -332,7 +333,7 @@ export class EventFeaturePlugin implements SkillFeature {
 			const host = this.getHost()
 
 			if (!host) {
-				throw new SpruceError({
+				throw new SchemaError({
 					code: 'MISSING_PARAMETERS',
 					parameters: ['env.HOST'],
 					friendlyMessage: `Stop! I need you to run \`spruce set.remote\` so I know where to connect! Or, you can set HOST in the env directly.`,
@@ -350,7 +351,7 @@ export class EventFeaturePlugin implements SkillFeature {
 			const { client } = await this.apiClientPromise
 
 			return client
-		} catch (err) {
+		} catch (err: any) {
 			this.apiClientPromise = undefined
 			throw err
 		}
@@ -398,7 +399,7 @@ export class EventFeaturePlugin implements SkillFeature {
 				currentSkill = skill
 
 				this.log.info(`Authenticated as ${currentSkill?.slug}.`)
-			} catch (err) {
+			} catch (err: any) {
 				await client.disconnect()
 				this.apiClientPromise = undefined
 				throw err