+++ packages/spruce-store-plugin/src/errors/SpruceError.ts
@@ -14,11 +14,6 @@ export default class SpruceError extends BaseSpruceError<ErrorOptions> {
 				}
 				break
 
-			case 'MISSING_PARAMETERS':
-				message = 'Looks like some parameters are missing: \n\n'
-				message += options.parameters.join('\n')
-				break
-
 			default:
 				message = super.friendlyMessage()
 		}